Fix several sign-conversion warnings in interface headers
This includes can trigger warnings when -Wsign-conversion is set. All
conversions are fine, so silence them by explicit casts.

+++ include/copperplate/clockobj.h
@@ -143,7 +143,7 @@ void clockobj_ns_to_timespec(ticks_t ns, struct timespec *ts)
 	unsigned long rem;
 
 	ts->tv_sec = (time_t)cobalt_divrem_billion(ns, &rem);
-	ts->tv_nsec = rem;
+	ts->tv_nsec = (long)rem;
 }
 
 #else /* CONFIG_XENO_MERCURY */
